+++ packages/icons/src/myst.tsx
@@ -34,7 +34,7 @@ export function MystLogo({
   );
 }
 
-export function MadeWithMyst({ url = 'https://myst-tools.org/made-with-myst' }: { url?: string }) {
+export function MadeWithMyst({ url = 'https://mystmd.org/made-with-myst' }: { url?: string }) {
   return (
     <a
       className="flex w-fit mx-auto text-gray-700 hover:text-blue-700 dark:text-gray-200 dark:hover:text-blue-400"