+++ packages/jupyter/src/jupyter.tsx
@@ -5,27 +5,34 @@ import type { MinifiedOutput } from 'nbtx';
 import { convertToIOutputs } from 'nbtx';
 import { fetchAndEncodeOutputImages } from './convertImages';
 import type { ThebeCore } from 'thebe-core';
-import { useThebeCore } from 'thebe-react';
-import { useCellRef, useCellRefRegistry, useNotebookCellExecution } from './providers';
+import { useNotebookCellExecution } from './providers';
 import { SourceFileKind } from 'myst-common';
 import { useXRefState } from '@myst-theme/providers';
+import { useThebeLoader } from 'thebe-react';
 
 function ActiveOutputRenderer({ id, data }: { id: string; data: IOutput[] }) {
-  const ref = useCellRef(id);
   const exec = useNotebookCellExecution(id);
+  const ref = useRef<HTMLDivElement>(null);
 
   useEffect(() => {
-    if (!ref?.el || !exec?.cell) return;
+    if (!ref.current || !exec?.cell) {
+      console.debug(`No cell ref available for cell ${exec?.cell?.id}`);
+      return;
+    }
+    if (exec.cell.isAttachedToDOM) {
+      console.debug(`Cell ${exec.cell.id} already attached to DOM`);
+      return;
+    }
     console.debug(`Attaching cell ${exec.cell.id} to DOM at:`, {
-      el: ref.el,
-      connected: ref.el.isConnected,
+      el: ref.current,
+      connected: ref.current.isConnected,
       data,
     });
-    exec.cell.attachToDOM(ref.el);
+    exec.cell.attachToDOM(ref.current);
     exec.cell.render(data);
-  }, [ref?.el, exec?.cell]);
+  }, [ref?.current, exec?.cell]);
 
-  return null;
+  return <div ref={ref} data-thebe-active-ref="true" className="relative" />;
 }
 
 function PassiveOutputRenderer({
@@ -39,67 +46,66 @@ function PassiveOutputRenderer({
   core: ThebeCore;
   kind: SourceFileKind;
 }) {
-  const [cell] = useState(new core.PassiveCellRenderer(id, undefined, undefined));
+  const cell = useRef(new core.PassiveCellRenderer(id, undefined, undefined));
   const ref = useRef<HTMLDivElement>(null);
-  useEffect(() => {
-    cell.render(data, kind === SourceFileKind.Article);
-  }, [data, cell]);
+
   useEffect(() => {
     if (!ref.current) return;
-    cell.attachToDOM(ref.current, true);
+    cell.current.attachToDOM(ref.current, true);
+    cell.current.render(data, kind === SourceFileKind.Article);
   }, [ref]);
+
   return <div ref={ref} data-thebe-passive-ref="true" />;
 }
 
-const MemoPassiveOutputRenderer = React.memo(PassiveOutputRenderer);
+export const JupyterOutputs = React.memo(
+  ({ id, outputs }: { id: string; outputs: MinifiedOutput[] }) => {
+    const { core, load } = useThebeLoader();
+    const { inCrossRef } = useXRefState();
+    const { data, error } = useFetchAnyTruncatedContent(outputs);
+    const [fullOutputs, setFullOutputs] = useState<IOutput[] | null>(null);
+    const exec = useNotebookCellExecution(id);
 
-export const JupyterOutputs = ({ id, outputs }: { id: string; outputs: MinifiedOutput[] }) => {
-  const { core, load } = useThebeCore();
-  const { inCrossRef } = useXRefState();
-  const { data, error } = useFetchAnyTruncatedContent(outputs);
-  const [loaded, setLoaded] = useState(false);
-  const [fullOutputs, setFullOutputs] = useState<IOutput[] | null>(null);
-  const registry = useCellRefRegistry();
-  const exec = useNotebookCellExecution(id);
+    useEffect(() => {
+      if (core) return;
+      load();
+    }, [core, load]);
 
-  useEffect(() => {
-    if (core) return;
-    load();
-  }, [core, load]);
+    useEffect(() => {
+      if (!data || fullOutputs != null) return;
 
-  useEffect(() => {
-    if (!data || loaded || fullOutputs != null) return;
-    setLoaded(true);
-    fetchAndEncodeOutputImages(data).then((out) => {
-      const compactOutputs = convertToIOutputs(out, {});
-      setFullOutputs(compactOutputs);
-    });
-  }, [id, data, fullOutputs]);
+      fetchAndEncodeOutputImages(data).then((out) => {
+        const compactOutputs = convertToIOutputs(out, {});
+        setFullOutputs(compactOutputs);
+      });
+    }, [id, data, fullOutputs]);
 
-  if (error) {
-    return <div className="text-red-500">Error rendering output: {error.message}</div>;
-  }
+    if (error) {
+      console.error(error);
+      return <div className="text-red-500">Error rendering output: {error.message}</div>;
+    }
 
-  if (!inCrossRef && registry && exec?.cell) {
+    if (!inCrossRef && exec?.ready) {
+      return (
+        <div>
+          {!fullOutputs && <div className="p-2.5">Loading...</div>}
+          {fullOutputs && <ActiveOutputRenderer id={id} data={fullOutputs} />}
+        </div>
+      );
+    }
+
     return (
-      <div ref={registry?.register(id)} data-thebe-active-ref="true">
+      <div>
         {!fullOutputs && <div className="p-2.5">Loading...</div>}
-        {fullOutputs && <ActiveOutputRenderer id={id} data={fullOutputs} />}
+        {fullOutputs && core && (
+          <PassiveOutputRenderer
+            id={id}
+            data={fullOutputs}
+            core={core}
+            kind={exec?.kind ?? SourceFileKind.Notebook}
+          ></PassiveOutputRenderer>
+        )}
       </div>
     );
-  }
-
-  return (
-    <>
-      {!fullOutputs && <div className="p-2.5">Loading...</div>}
-      {fullOutputs && core && (
-        <MemoPassiveOutputRenderer
-          id={id}
-          data={fullOutputs}
-          core={core}
-          kind={exec?.kind ?? SourceFileKind.Notebook}
-        ></MemoPassiveOutputRenderer>
-      )}
-    </>
-  );
-};
+  },
+);